requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.1
django-allauth==0.24.1
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-ckeditor==5.0.3
django-extensions==1.6.1
django-image-cropper==0.3.0
django-image-cropping==1.0.3
django-simple-captcha==0.5.1
django-suit==0.2.16
django-suit-redactor==0.0.3
easy-thumbnails==2.3
oauthlib==1.0.3
Pillow==3.1.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
static3==0.6.1
wheel==0.24.0
